pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "so2x-system"
version = "0.1.0"
description = "Thin orchestration and evolution layer for Superpowers-style execution flows"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "kim" }]
license = { text = "MIT" }
dependencies = ["PyYAML>=6.0"]

[project.scripts]
so2x-system = "so2x_system.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
